feat: merge into one VIP the risk parameters commands
